Type the finite Number constants as float64 so MAX_SAFE_INTEGER lowers to a float64
Number.MAX_SAFE_INTEGER and Number.MIN_SAFE_INTEGER were written as bare integer
literals, so an untyped integer constant defaulted to int in a := binding. A
const bound to Number.MAX_SAFE_INTEGER then refused to flow into a float64
parameter (value.Some[float64], the Duration field builders), failing the Go
build on a cluster of Temporal cases. Every JS Number is a double, so type all
five finite Number constants float64. The two internal Go range checks in the
goimport bridge convert their int64/uint64 operand to compare against the
now-float64 bound; the check is unchanged in meaning.

+++ pkg/goimport/bridge/bridge.go
@@ -40,7 +40,7 @@ func StringFromGo(s string) value.BStr { return value.FromGoString(s) }
 // small Go integer types always fit, so they convert with a plain float64() in the
 // emitted code and never call here.
 func Int64ToNumber(n int64) float64 {
-	if n > value.NumberMaxSafeInteger || n < value.NumberMinSafeInteger {
+	if float64(n) > value.NumberMaxSafeInteger || float64(n) < value.NumberMinSafeInteger {
 		panic(RangeError{Message: "go: integer result out of Number.MAX_SAFE_INTEGER range"})
 	}
 	return float64(n)
@@ -50,7 +50,7 @@ func Int64ToNumber(n int64) float64 {
 // safe-integer check as Int64ToNumber; the lower bound is zero, so only the upper
 // bound can trip.
 func Uint64ToNumber(n uint64) float64 {
-	if n > value.NumberMaxSafeInteger {
+	if float64(n) > value.NumberMaxSafeInteger {
 		panic(RangeError{Message: "go: unsigned integer result out of Number.MAX_SAFE_INTEGER range"})
 	}
 	return float64(n)
